Fix issues with tracker config and network error handling, update migration guide (#57)
* Rename public constants 'TrackerConfig' to 'Config' to match previous Media name.

* In migration-guide, add section for changed Java package name.

* Remove code to clean tracker config as it is unnecessary

* Remove optional unwrapping for 'tracker' and 'player' in MediaAnalyticsProvider as neither is Optional

* Update createTrackerWithConfig unit tests with updated expected config parameters

* Fix MediaSession.handleErrorResponse to cast error status code to int instead of assuming long type

* Add 'ignoreCase' label when calling String.equals

* Rename MediaConstants.Config back to MediaConstants.TrackerConfig

* Use 'TrackerConfig' name in docs

* Undo changes to MediaTestApp

* correct spelling in migration-guide

* Use tracker config variable in test app when creating media tracker

+++ code/edgemedia/src/main/java/com/adobe/marketing/mobile/edge/media/internal/MediaInternalConstants.java
@@ -37,7 +37,7 @@ final class MediaInternalConstants {
         static final String REQUEST_EVENT_ID = "requestEventId";
         static final String PAYLOAD = "payload";
         static final String SESSION_ID = "sessionId";
-        static final long ERROR_CODE_400 = 400L;
+        static final int ERROR_CODE_400 = 400;
         static final String ERROR_TYPE_VA_EDGE_400 =
                 "https://ns.adobe.com/aep/errors/va-edge-0400-400";
